Allow NULL toEscape in jsonsl_util_unescape_ex
Passing NULL for toEscape is now supported and means 'unescape
everything'. This is simpler for the caller, and also runs
noticeably faster since the array doesn't have to be checked
on every escape char.

I also simplified the flow of control, removing the in_escape flag,
to further speed up the loop.